move getComponentStateFromSpec to environments package
